HomeWork3

+++ src/test/java/com/samsung/dao/QuestionDaoTest.java
@@ -1,42 +1,71 @@
 package com.samsung.dao;
 
-import com.opencsv.CSVWriter;
-import com.samsung.domain.Question;
 
-import org.junit.Test;
+import com.samsung.domain.Question;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.io.File;
 import java.io.FileWriter;
+import java.io.IOException;
 import java.io.PrintWriter;
 import java.util.ArrayList;
+import java.util.List;
 
+@DisplayName("Класс QuestionDao")
 public class QuestionDaoTest {
+    private final String csv = "dataFindAllTest.csv";
+    private QuestionDaoSimple dao;
 
+    @BeforeEach
+    public void setUp(){
+        dao = new QuestionDaoSimple(csv);
+    }
+
+    @DisplayName("Должен возвращать все вопросы")
     @Test
-    public void findAllTest(){
+    public void shouldFindAllQuestions(){
         try {
-            String csv = "dataFindAllTest.csv";
+        if(!new File(csv).exists()) new File(csv).createNewFile();
             PrintWriter writer = new PrintWriter(new FileWriter(csv));
             String csvParams = "1;How much is 5 * 5?;25\n" +
                     "2;How much is 4 + 3?;7\n" +
                     "3;How much is 1 - 1?;0";
             writer.write(csvParams);
             writer.close();
-            QuestionDaoSimple questionDaoSimple =
-                    new QuestionDaoSimple("dataFindAllTest.csv");
-            ArrayList<Question> questions = new ArrayList<>() {
-            };
-            questions.add(new Question(1, "How much is 5 * 5?", "25"));
-            questions.add(new Question(2, "How much is 4 + 3?", "7"));
-            questions.add(new Question(3, "How much is 1 - 1?", "0"));
-
-            ArrayList<Question> questionsFromMethod =
-                    (ArrayList<Question>) questionDaoSimple.findAll();
-            File file = new File("dataFindAllTest.csv");
-            file.delete();
-            assert(questions.equals(questionsFromMethod));
-        }catch (Exception e){
+        } catch (IOException e) {
             e.printStackTrace();
         }
+        List<Question> questionList = new ArrayList<>();
+        questionList.add(new Question(1, "How much is 5 * 5?", "25"));
+        questionList.add(new Question(2, "How much is 4 + 3?", "7"));
+        questionList.add(new Question(3, "How much is 1 - 1?", "0"));
+        assertEquals(dao.findAll(), questionList);
+    }
+
+    @DisplayName("Должен возвращать вопрос по id")
+    @Test
+    public void shouldFindQuestionById(){
+        Question question = new Question(1, "How much is 5 * 5?", "25");
+        assertEquals(question, dao.findById(1));
+    }
+
+    @DisplayName("Должен добавлять вопрос")
+    @Test
+    public void shouldAddQuestion(){
+        int sizeBefore = dao.findAll().size();
+        Question question = new Question(4, "How much is 5 * 9?", "45");
+        dao.save(question);
+        assertEquals(sizeBefore + 1, dao.findAll().size());
+    }
+
+    @DisplayName("Должен удалять вопрос по id")
+    @Test
+    public void shouldDeleteQuestionById(){
+        int sizeBefore = dao.findAll().size();
+        dao.deleteById(1);
+        assertEquals(sizeBefore - 1, dao.findAll().size());
     }
 }
